monitoring-server/requirements.txt
alembic==1.7.6
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.1
enum-compat==0.0.3
Flask==2.0.3
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
lgpio==0.0.0.2
Mako==1.1.6
MarkupSafe==2.1.0
paho-mqtt==1.6.1
protobuf==3.20.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.9
Werkzeug==2.0.3
yandexcloud==0.159.0
ydb==2.3.1
